The big ass shader update (WIP)
- Fixed some warnings
- Fixed glew initialization
- Added the material class
- Added shader loader
- Added 2d mesh class
- Added 2d mesh render component
- Added a material manager
- Added vertex helpers
- Added mesh helpers

+++ SpartanFramework/MathHelpers.h
@@ -196,7 +196,7 @@ namespace Math
 	inline Vector2 LengthDir(float length, float dir)
 	{
 		Vector2 result = Vector2::Zero();
-		float radians = dir / 180.0f * M_PI;
+		float radians = dir / 180.0f * (float)M_PI;
 		result.x = cos(radians) * length;
 		result.y = sin(radians) * length;
 		return result;
